fix(snapshot): ignore missing files on abort
Use FileUtil#deleteFolderIfExists(Path) on transient snapshot abort.
This aligns the abort behavior for all types of snapshots to ignore
missing files, as when deleting/aborting snapshots, it's fine if some
files were previously already deleted (for example due to rollback in
the store) - we still end up with the desired behavior of all the files
being removed.