fix(engine): clean up eventScope
Instead of deleting the event scope on deploying a new process, we
remove the event scope either when we have triggered the ProcessEvent
or when we canceled the Timer. Otherwise we have problems on triggering
the timer and deploying a new version concurrently, this caused problems
since the event scope doesn't exist anymore on activating the process
instance.