refactor(broker): reduce log level to warn
Reduces the log level of a failure to apply a snapshot chunk to a
warning instead of an error, as this is most likely recoverable if it is
an IOException. This also aligns it with our logging guidelines.

(cherry picked from commit c02af7f4cac35313c77717d5b917b17a225a6667)