refactor(test-util): rework cutOffPositions
It was not really clear what are these cut off positions, the code made
it a bit hard to understand. This commit tries to clean that a bit up
with introducing a new class ExecutionBoundaries and more java doc,
which should explain a bit more the purpose of it. Furthermore the logic
itself was simplified instead of iterarting over it again or calculating
certain values we just store all boundaries in a list, since we don't
expect to have much of them.