fix(broker): fixes race condition when closing
Fixes a race condition between removing the component from the critical
monitor and transitioning to inactive (which sets the status to
UNHEALTHY). This is because both calls use `actor#run`, which enqueues
jobs at the front of the fast lane queue. As such, calling
`removeComponent` and then `onComponentFailure` (which is called via
transitionToInactive) means the second call may be executed first.

By moving the remove component when the close is requested, we avoid
this race condition since the monitor and the partition share the same
actor.

(cherry picked from commit e5ca7ceb30ba723b683769d27e9304c054636ad6)