refactor(clients): wrap ObjectMapper instead of extending it
I previously accepted a PR where we extend a library defined type,
ObjectMapper, instead of providing an adapter for our interface.
Extending a class we have no control over for something like this is a
bad idea, as you may inadvertently start using inherited features from
the new class where you shouldn't (since it's not meant to be a drop in
replacement of ObjectMapper). Instead, we now wrap an existing one and
use it where we need, and only expose the methods we need to.