fix(journal): delete unacked entries on partially written descriptor
When corruption is detected while loading the segment's descriptor, we compare the previous
segment's index to the metastore's last written index to determine if we can
delete it and the subsequent segments. In the "common" case of a partial write, the index
was not written to metastore so it's safe to simply delete.