fix(engine): use already defined loop counter
After resolving an incident the failed command is processed again. If
that command is Element:Activate for a multi-instance marked element,
then the loop counter has already been incremented and would be
incremented again. To make sure we continue at the correct loop index,
the loop counter (which is used as index) can actually be used directly
from the child instance, since it was already set on the child instance
activating.