fix(broker): use correct term and role when reacting to install failures
The role transitions are enqueued in ZeebePartition actor. When there are more than one role transition happened one after the other in a short span, ZeebePartition doesn't know which installation has failed. Previously it was reading the term and role from the RaftPartition. However, since the role transitions are asynchronous the role and term on the RaftPartition might be already refering to a newer term and role.

To fix this, now PartitionContext keep track of current term and role which is updated when the corresponding install steps has started. On install failure, ZeebePartition can read the term and role from the context to determine how to react to the failure.