refactor(engine): merge transitionToCompleted methods
There should not be a reason for a client to call the variant of these
methods that does not inform the parent. The only case is specifically
for processes, but that can be checked inside the merged method.

(cherry picked from commit 66555bd1cfac0b7e80dd2374c27af52e1bc70445)