feat(zb-db): allow manual compaction for rocksb state
Using manual compaction of rocksdb, we can compact files from each columnfamily into a target outputlevel. In our tests we found that if we set a lower level as output level, and if there files in higher level, then rocksdb compact them to the higher level. This is acceptable for our case.