feat(broker): force manual compaction of state on recovery amd expose feature flag to turn it on/off
Forced manual compaction is necessary in some cases where rocksdb is not compacting the state even though auto-compaction is enabled resulting in thousands of file in the state. Using manual compaction of rocksdb, we can compact files from each columnfamily into a target outputlevel.

Since this forced compaction is only necessary in some cases we also expose a feature flag to turn it on/off. To prevent unnecessary compaction, we trigger the compaction only if the number of files is above a threshold.

In our tests we found that if we set a lower level as output level, and if there files in higher level, then rocksdb compact them to the higher level. This is acceptable for our case. Hence as default we set the output level to 1.