fix(broker): delete runtime folder when closing state
When closing the state, we may leave behind a copy of the database,
which is only deleted the next time we open the state. This causes
unnecessary data usage as that copy is anyway thrown away on the next
open. With this fix, we now delete it always when opening and when
closing.