fix(engine): dont complete child when parent fails
When a child completes (complete command is being processed), it may
inform the parent (flowscope) of it's completion (specifically when the
child is at the end of the execution path within it's flowscope).

The parent processor can perform some logic when it is informed of the
completion of the child (specifically when the child is completing and
when the child has completed). In some cases, the parent processor is
allowed to fail executing this logic, which should result in an incident
on the child element (forcing it to complete again when the incident
resolves), and it should not write that the child has completed
(otherwise, the child element would've been removed from state making it
impossible to resolve this incident).

This commit allows parent processors to fail the
beforeExecutionPathCompleted by returning either a failure or success.
All child processors need to deal with this accordingly.

In some specific cases (e.g. ExclusiveGateway), the onComplete method is
not implemented, so it would not be possible to resolve the incident.
However, at this time the exclusive gateway would never be the end of an
execution path for a parent element that could fail the method
beforeExecutionPathCompleted. If this changes in the future, and an
incident is created by these processors, then we can implement this
onComplete method at that time to allow the incidents to be resolved in
a future version.

(cherry picked from commit f36e75a35919631eba4eecd7d340249fd9de93c0)