test(atomix): fix flaky swim test
It is possible that member1 detects member2 as a suspect before it is removed. This can happend if member1 probes old member2 before member2 with new version communicates with member1. Hence the check is updated to wait until the MEMBER_ADDED event occurs instead of immediately excpecting the MEMBER_ADDED event.