fix(clients/java): log RESOURCE_EXHAUSTED on trace
Logging the RESOURCE_EXHAUSTED on warning level caused some confusion in
the community and it produced a lot of noise in our logs. In most cases
it is expected if the broker is overloaded or is quite busy, our backoff
strategy will handle this case. For these cases we don't need to see
this exception.

If there is an issue with activation and we need to investigate that we
can lower the log level to trace to see the exception again.