fix(util): serialize datasize to lower units
Previously datasize was always serialized to MB, that means any value less than 1MB will be serialized to 0MB. This commit fixes it by serializing into lower units. All higher values will be always serialized to MB. One problem with this approach is that the values are rounded down. 1.5MB will be serialized to 1MB. This is ok because this serializer is only used to print out the configuration parameters.