test(engine): should collect output when incident resolved
Previously, it was possible that an incident was created for an element
that was already removed, which led to multiple problems. One of those
problems is that it never allowed for correctly informing the container
processor that an element completed as part of resolving that incident.
Mostly, because the incident couldn't really be resolved, but also
because the container could still continue if all other children would
finish their processing, which is a very strange state to be in.

This test verifies that a container processor (multi-instance) is
informed of the child completing (which results in the collection of the
output in the result variable), when an incident is resolved that was
created by the failure of that parent.