test(update-tests): add rolling update test
Bring back the rolling update tests, with minor updates for 1.0. It was
previously decided that we would drop support for it, but this has been
changed to making it optional. We should strive to allow rolling
updates, but if need be, it can be broken between minor versions.
However, it's important that we're aware of it so we can advise users
correctly on how to udpate.