fix(engine): dont trigger event subproc for boundary events
In case of the #7097 the boundary event was tried to be cast to a start
event because its flow scope was the event sub process. However, that
part of the logic should only be executed for start events, not boundary
events. By restricting the if statement, boundary events pass to the
next if branch and can correctly terminate the active scope.