refactor(broker): simplify jar class loader
Simplifies the JAR class loader behavior by shifting some responsibility
onto the exporter authors. Previously, some packages were directly
loaded from the system class loader and not from the JAR itself,
presumably to avoid potential dependency/version issues if a JAR would
package, say, the exporter API as well.

I consider this unnecessary complexity, as not only is it unclear why
only these packages are checked, but it's also unclear to exporter
authors that this behavior exists, and the onus should be on them to
only package the libraries they need and exclude everything else.