refactor(util): extract CheckedRunnable
Extracts CheckedRunnable interface out of the broker's bootstrap and
consolidate it with util's ThrowingRunnable, as they both serve the same
purpose, and the broker's bootstrap package is too specific to indicate
an intent for reuse.

Additionally fixes some warnings in LogUtil by adding a private
constructor for a utility class to avoid instantiation.