fix(snapshots): remove itself from pending list after snapshot is persisted
FilebasedSnapshotStore keeps track of a list of pending snapshots which will be used to deleted them. When a pending snapshot is committed, it was not deleted from the list. Later when snapshot store purges all pending snapshots, it tries to delete also this non existing pending snapshot which caused unnecessary warning in the log message. This commit fixes it by removing the pending snapshot from the list after a snapshot is successfully committed.