refactor(engine): pass keygenerator to processor
The KeyGenerator will be required by the ResolveIncidentProcessor to
create a key for the new Incident:Created event.

Keygenerator is a state mutator that is allowed to be used in
processors. Since the ResolveIncidentProcessor no longer has access to
MutableZeebeState, the KeyGenerator needs to be passed in.

(cherry picked from commit dce77f6a44bdfe49219ea52ac1e52435b33b3e09)