ci(github): only trigger backport on merged pulls
Closing a pull request would also trigger the backport action, which
would then comment on that pull request that it was not merged. We only
want that behavior when you explicitly try to trigger the action with
the /backport comment, so that the bot replies (as conversational ui).

This commit uses yaml folding to nicely layout the different conditions
when to trigger the action and restricts the 'pull_request' events to
only merged pull requests.