ci(github): sync backport with action version 0.0.4
Previously the master version was used to easily fix problems with the
backport action in the zeebe repo, but it has been worker rather stable
recently. It thus makes sense to start using a specific version and
update more explicitly when a new version is available.

I've simply copied the recommended 0.0.4 configuration.

We could consider enabling dependabot's update features for github
actions.