test(exporters): fix ES container memory
It seems the default configuration for the ElasticsearchContainer is to
use all the heap available, even as initial heap. This is unnecessary
for our tests and can cause issues with other tests running in parallel
on the same machine. This commit fixes the max heap size to 1Gb and the
max direct memory to 1Gb as well.