test(qa): fix ES memory usage
Elasticsearch will automatically grab all available memory as its
initial and max heap size on start up. When starting it as a container
with no particular limits, it will then use up all the host's heap,
which causes problems for other processes running on the same node.

This was already fixed in the ES exporter tests, but I missed this case
here.