fix(monitor): increase role change granularity
In some cases, the role change happens so fast and changes back to the
previous state (in less than 15s). When that happens, grafana was unable
to show the election cycle. By reducing the query interval for the
datasource we can increase the granularity and show these role changes.
The interval is only changed for the role change panel, to not impact
performance.