fix(qa): remove useless instance from test
The second process instance that creates a job that can be activated
does not really do anything for this test. The test is to check that a
job is ignored when it's variables are too large to fit in a message.

We can simply create the process instance, set the variables and try to
activate it. It should return no jobs. In order to return immediately we
do need to disable long polling.

(cherry picked from commit 27625f5daaa81f9edab45d265cb30d5c3aa7c1af)