fix(engine): fail resolving missing incident
In case the failed command can no longer be found for the incident to
resolve, we should not be able to resolve the incident. Instead, the
resolve command should be rejected. This already happened, but it would
simultaneously also write incident resolved (which seems conflicting).

To match behavior of other unrecoverable parts in the engine, we just
throw an exception during the processing, which stops any follow-up
events from being written and rejects the command.

(cherry picked from commit 2658a5fc07cbbb9180d8c5b4912dd5e995836bf3)