build(ci): scale down dind only for go tests
Remove integration of dind sibling for the maven unit test container,
and scale down the resources of the dind sibling. It's still present in
the distribution template because the Go integration tests rely on it,
but we still want to make sure no unit tests on the Java side will work
with it. Plus, it frees up more resources for us to parallelize our
unit tests.