build(parent): ensure revapi dependency is present offline
Even though we skip the revapi during integration testing, we still need
to ensure the org.revapi:revapi-java dependency (which the plugin
requires) is present offline, as otherwise the plugin will fail to load
and the execution is not skipped but rather failed. Additionally, since
we now enforce it so it's downloaded, we need to make sure it's ignored
even though it's unused.

(cherry picked from commit bde52bd383b0518b0ac4617f4ddbdb071b9d9369)