fix(engine): complete open later
The StreamProcessor was marked as open to earlier, before the Reprocessing was done. This can cause issues like opening the CommandAPI too early, which fill already the command queue and return a lot of resources exhausted responses to the clients. It is not useful to accept already commands, if we are not able to process them

(cherry picked from commit 9d92be58ce6477115a10e09fbee0e459480d1307)