build(parent): skip jacoco when unit tests skipped
We are only interested in the test coverage for unit tests, not for
integration tests. It is anyways hard disabled to run in the qa module.
For the other modules however, we can simply enable/disable it based on
whether or not we run the unit tests, or rather skip them

(cherry picked from commit 8863c93969cce9d2e3cfbc88e41eaafa28c5df5d)