build(Dockerfile): introduce Maven builder stage
Introduces a new stage which acts as an alternative to the DISTBALL
builder. By default, the entrypoint of the file is still the DISTBALL
builder and works as before. However, if you specify the build argument
BASE=maven, then the project is copied into a maven container and built
from scratch. This is time consuming, as all dependencies must be
downloaded again, but works. This prepares the image to be built by
Testcontainers on demand when need be.

A .dockerignore is also introduced to keep the build context small, e.g.
excluding target/packaged JARs, but also to avoid having to rebuild due
to cache invalidation when modifying test classes, for example.