feat(test-util): optionally build local Docker image for tests
Adds a new feature toggle to optionally build the local Docker image
when running a test which wants to start a container with it.

This only works if all tests always use
ZeebeTestContainerDefaults#defaultTestImage(), of course.

By default, whenever the image is called, it will build a new, fresh
version of camunda/zeebe:current-test, but leveraging the Docker build
cache. It will only build the image once per run - so if you run
multiple tests in the same JVM run, it will not rebuild the image.

This feature is disabled for CI as we already build the image once in a
specific stage, and we know it's fresh as we just rebuilt it, so there's
no need to build it when running tests. It's only really useful to build
it on the fly when running tests via IntelliJ.

If you wish to disable this and build the image yourself, you can of
course set that environment variable globally, or use a dotenv plugin in
IntelliJ to load environment variables from a dotenv file.