refactor(atomix): improve logging on election timeout
Improves logging when an election occurs due to missing heartbeats. This
fixes the issue of having very high "time since last heartbeat" due when
there are no heartbeats, and it also shows that there is no known leader
instead of just showing null.

Additionally, adds logging when sending poll requests to log if we
transition directly to candidate or, when not, to which members we are
sending poll requests.

(cherry picked from commit 59b264d592046d05cfabefa649c003687ed5c0b1)