refactor(raft): track hearbeat time in raft context
Previously heartbeat time is tracked in the follower role. If a node transition between follower and candidate and back to follower, the last heartbeat time is reset. This is not ideal as the actual last heartbeat was received before the node transitioned to candidate role. The last heartbeat time will be used to calculate election latency, hence the actual time when the last heartbeat was received is important.