fix(atomix): rewrite request timeout logic
Rewrites the request timeout logic in NettyMessagingService by
scheduling the request timeout to include the complete request
lifecycle, including creating the connection, performing the handshake,
sending the request, receiving the response, etc. Previously, it was
scheduled only after performing the handshake, which means if anything
took too long or went wrong before, the contract of the method was
broken.

The number of threads for the time out executor was also lowered from 4
to 1, as the executor shouldn't be doing anything but failing futures
(so it's not expected to be doing much).