fix(atomix): report channel close event
Ensure we report the channel close event by failing the future if the
future has not yet been completed. Previously, if the future was not
completed (e.g. the channel was closed mid-way through the handshake),
we were not notified of the channel closing, and the future remained
open forever.

(cherry picked from commit 3c2772fc0d597e188c09322f567e81faff85220f)