refactor(broker): hide Atomix class behind a decorator
This change removes all access to the Atomix class in production code. Instead, a new class ClusterServices is introduced that delegates all access to the Atomix class. The benefit of having this indirection is that ClusterServices has only a minimal interface, whereas Atomix has a large interface. For test classes a reference to the underlying Atomix class is still provided.