docs(benchmarks): update with OAuth setup
Updates the documentation regarding how to setup one of our benchmark
clusters, specifically how to turn on Github OAuth2 authentication for
Grafana. This serves for further clusters that will need to be setup in
the future, but also as documentation of the current settings deployed.

It removes the need for a shared admin password, and lets us use normal
users, such that we can safely share Grafana URLs publicly.