build(ci): track each flaky test separately
Moves the flaky test detection to the complete end of the pipeline,
avoiding issues with the parallel stage and figuring out in which agent
the flaky test analysis is ran.

Additionally, track the build status of every flaky test separately so
we can track each case by itself. This will allow us to get statistics
on which tests is "flakiest", and which have stopped happening.