refactor(broker): bring notification of listener closer to transition
The listeners are now private to the transition context. This is an intermediate step.
I would like to notify the listeners as last step in the transition.
Current problem is that the new term is not readily available to the transition steps.

This change also has the benefit that we can remove the LogStream form the PartitionContext interface