build(parent): fix flaky test extraction
Modifies the flaky test extractor profile such that it's only activated
manually. The reasoning here is that this is only used for CI and not
for local development, so it's fine to have to manually enable it.

This also fixes the usage of the plugin, which was not working with
failsafe since it was not looking in the right directory. The profile
now has a property which it uses to specify whether to look into
surefire-reports or failsafe-reports. It was done this way because
normally the plugin expects a full directory, but here we don't know the
project's directory, just that in there it should look up
failsafe/surefire reports. So instead of specifying it on all projects,
we just have to set one property before running maven.

One downside is the plugin has to run in the post-integration-test
phase, so we now unnecessarily package the application even for unit
tests.