build(parent): remove flaky test profile
Removes the flaky test extractor profile which was automatically added
whenever tests were run. We will from now on run the plugin explicitly
where needed (e.g. CI). The reason is that we only really need this for
CI - typically developers won't look at the surefire reports on their
local machine, but rather use the IDE to run the tests. Additionally,
since it binds to the post-integration-test phase, it meant we had to
run up to there (including packaging) even when only running the unit
tests, which was a waste of time.

Additionally ensure the plugin is available even when going offline.