refactor(broker): remove dependency of partition components on CommandApiService
This removes the dependency of the partition components on the CommandAPIService. I first thought about defining an interface with the methods of CommandAPIService that the partition depends upon. But then I could not find a good name for that interface. So it is now two small things that the partition depends upon.