test(engine): start processor without await opening feature
This is necessary for some tests, where we cause to fail the stream
processor opening. For example where we expect a endless loop in
reprocessing or throw an exception on a RecoveryListener. In these test
we want to verify whether the processor goes into a fail state and is
marked as unhealthy