refactor(broker): do not keep the reader opened in AtomixRecordEntrySupplier
Sonarlint check failed because the reader was not created with try-with-resources. This commit fixes it by moving the reader creation into AtomixRecordEntrySupplierImpl. This class is used only by the snapshotting process. That means we don't have to keep the reader open for ever. Instead we can create reader only when we want to read the previousIndexedEntry.