feat(atomix): introduce new committed entry listener
The new comitted entry listener is called only in a Leader role, and
serves only one purpose, to make it possible to have access of the last
committed entry (without an explicit reader).

This is useful for leader services to resolve the last committed position without
creating a mapping nor having a separate reader. The Snasphot director
of a leader will use that to make sure that the snapshots are valid.