refactor(logstreams): add commitListener interface to LogStorage
Using this commitListener, all consumers of LogStorage can be notified of new records committed, even if the record was not writter by the consumer itself. Previously only logstreams on the leader node was able to get notified of the commit via the appender. In future we want to run StreamProcessor on followers as well. Then the logstreams running on follower can get notified of new records using this listener.

The commitListeners in AtomixLogStorage must be a threadsafe a collection, as there can be concurrent add, remove or notify operations on the listeners.