feat(util): add Either collector
Adds a collector method to easily collect a Stream of Eithers into
either a Left or a Right. This is handy in case you are dealing with
multiple things that can cause errors. It simply collects all errors
together in a list if there were any, and otherwise it collects all
successes.