fix(broker): avoid race condition
The logStorage was concurrently manipulated by different actors, since
the step was installed for each partition. In order to avoid that the
storage is set in to the partition context, which is an object for each
partition.

I dont like to add more to the context, but it quickly solves the issue for now.
Alternative solution would be to have instead of one object for all partitions, separate
objects. Like defining only the classes for the Leader- and Follower-Steps.
Since we plan to refactor the bootstraping I don't think it makes sense to spent
more time on this.