feat(engine): add cont. replay mode
On the creation of the StreamProcessor we can set in which mode the
processor should run. If we set the CONTINUOUSLY_REPLAY flag it means
the ReplayStateMachine will run for ever. It will register on the logstream
for new records and replay events. The replay future is only completed
on a failure.