fix(journal): improves durability of segment creation
There is a known issue on Linux, depending on your file system, where
writing to a new file and forcing may guarantee its data blocks to be
present, but will not necessarily guarantee that that file is visible in
that directory after a crash. For that, it's also necessary to fsync the
parent directory as well.

To simplify things at segment creation, we also now map the file
immediately. This both creates the file and also gives us a buffer where
to write our descriptor, avoiding having to allocate that and then copy
it, avoiding having to open and close the file again as well.

The error handling is also improved, as we move away from Agrona's
IoUtil which hides these errors, and we handle them as much as possible
or at the very least enrich them with context.