fix(monitor): show leader per partition in Grafana
The Grafana panel 'Current Partition Leader' was rather strange to read.
It showed a table with rows for each of the brokers that are leader for
a partition and then in the columns it showed the name of the broker and
the partition id for which it is leader.

It makes way more sense to show the partition ids in the left column and
then the broker which is leader for that partition on the right of it.
This way it reads as: partition x has broker y as its leader.