refactor(broker): extract PartitionService from Atomix
This commit is a bit large, but there was just no way around it.

The starting point was the attempt to remove the partition group from Atomix.
This led to the removal of the partition service.
This led to simplifying the factory methods.
And then this lead to removing Atomix, AtomixBuilder and AtomixConfig altogether.
As a side effect, this also allowed some refactoring in the Broker. The SnapshotFactory is now only needed when the partitions are installed.