feat(protocol-jackson): introduce new record implementations
Introduces a new implementation of the protocol's Record, RecordValue,
and related interfaces. This new implementation is specifically to be
compatible with Jackson and allow (de)serialization to and from JSON, as
well as any format which Jackson natively supports.

The implementation classes are generated by the Immutables library,
based on intermediate abstract types. We opted against annotating the
protocol definition itself with the Jackson annotation, or reusing the
implementations, because this use case is for integration with external
systems. The protocol definition is purely an API, and the protocol-impl
module is the specific implementation (which is a mix of SBE and
MessagePack) for the engine, and is meant to be used purely internally.
The new implementation is meant to be used externally, primarily in
testing, but also for Java based applications which want to deserialized
exported records, e.g. Operate, Tasklist, Optimize.

So for the Record interface, and any other abstract type for which we
want to provide an implementation, a new Abstract class is added. This
class serves as our customization point where we can specify Jackson
annotations to control (de)serialization of these objects, notably how
to perform the polymorphic deserialization of the Record's value object,
or the Intent interface to its concrete enum. This adds a bit of
boilerplate, but it was considered preferrable to the alternatives:
either hand roll our own implementation (more boilerplate), or annotate
the types directly in the protocol module and break the module
boundaries. It would then become both an API and also include an
implementation, as well as (de)serialization directives, both of which
would then have to be maintained as public.

Interally, the polymorphic mapping is done via TypeIdResolver
implementations which rely on a central ValueTypes class, which maps
ValueType to the correct RecordValue implementation class and the
correct Intent enum. There are smoke tests to ensure that we map all
known value types.

Testing is done by integration via the exporter test harness, using the
DebugHttpExporter, which already exports all records as JSON.