feat(util): add .peek to Either
Peek can be very handy when you want to perform some actions on a right
either, but also need to return the either in the end. Peek will only be
executing the action if it is a right, so peek can be executed before
returning the either from your method and then later wanting to act on
the failure case (left). So this is especialy useful when failure
handling and success handling are separated in different
methods/classes/components