fix(util): remove peek from Either
We favor ifRight over peek, so we can just remove it

Peek does not help making code easier to understand. ifRight forces
callers to highlight when side effects are produced, but peek can be
hidden in the callchain. In addition, Stream.peek has a notation in its
documentation that it should only be used for debugging purposes.