feat(util): add Either.collectorFoldingLeft
Adds an additional collector that collects rights and folds lefts.

This can also be done fully functional using .reduce(identity,
accumulator, combinator), but that cannot be provided as easily as a
single Collector, because there is no Reducer interface.

Using a reducer would be more performant, because you can use Either's
flatMap to shortcut when it encounters the first left (flatMap returns
this for a left). It's also less readable because the flatMaps and maps
required are tough to understand. For now, a collector makes more sense.