refactor(engine): move specialized type check to separate method
The type check for message correlation keys is different from the
others. It's useful that it adds extra information about what is wrong.
In the future we might be able to align that further with the other type
checks. For now, this mapLeft is still necessary and to reduce the
indentations lets move it to another method