feat(journal): identify segments marked for deletion
The segments that are deleted will be first marked for deletion by renaming it. They will be deleted later.

The new segments created after marking a segment to be deleted, can have the same segment id. Hence when renaming the file for deletion, we have to ensure that the new name is unique. To handle that we added an incrementing index to the name. Hence if "journal-1.log" is deleted, the renamed file will be of the form "journal-1_1-deleted".