feat(journal): defer deleting segment files until the readers are closed
When the segments file is immediately deleted and the buffer is unmapped, there will be concurrent readers reading from the deleted segment. This can cause segmentation fault as it is trying to access an unmapped buffer. To prevent this, this commit defers the deletion of the segment file until all the readers are closed. The readers are already tracked by the segment.

To ensure that segment file defered for deletion are deleted eventually in case of crash failures, they are deleted on load.