feat(raft): raft notifies listeners when a snapshot replication is started and completed
A follower running stream processing can remove all consumers of log and reinstall once the snapshot is committed. This can be triggered using the listeners.

It can happen that the snapshot replication is cancelled after onSnapshotReplicationStarted() is invoked. It can be cancelled due to different reasons such as a new leader was elected which does not have this snapshot. Hence it starts sending new events instead of the snapshot. In this case it is important the listeners are notified, otherwise the follower never restarts streamprocessing. To ensure this the listeners are also invoked when the snapshot is aborted.