feat(journal): read return false when the segment is already closed
When the segment is closed, the readers should not read further to prevent reading dirty records. Instead hasNext returns false. If the readers want to read further, it should reset to a valid index first.
The expectation here is that when the journal is reset, the readers are eventually reset or recreated. This should be enforces by the application using the journal.