feat(broker): snapshot director can distinguish processing and replay mode
On replay mode, snapshot director doesnot wait for the the last event to commit because only committed events are replayed. Hence all the events included in the snapshot is already committed. So it is safe for the snapshot director to immediatly persiste a snapshot.
When running on replay mode (on the follower) the commit listeners are not available. So it is not possible to wait for the commit. This is why we need different logic for snapshot director running on the follower and the leader.