refactor(util): attempt at simplifying actor based bootstrap
A first attempt at simplifying using the actor scheduler as concurrency
control for the bootstrap process, i.e. orderly startup and shutdown.

One of the key things here is the constraint that the BootstrapProcess
is owned by an actor, and its startup/shutdown method can only be called
from within that actor. This is unfortunately not enforceable at compile
time, but only through programmer diligence. When this is the case, then
any future returned BootstrapStep can be awaited on from within the
initial calling actor without having to worry about callbacks and
complex scheduling. This is because doing onComplete of a future gets
your current actor (whatever it is) and ensures the callback you give is
executed within its context. In our case, this means the context of the
actor owning the BootstrapProcess instance. So we don't have to worry
about synchronization that much, and can focus on making sure the
process is observable and orderly.

I added an example in the tests to show case this, see the test package
startup.actor, with the Application class as an entry point.