test(qa): improve flaky DiskSpaceMonitoringFailOverTest
StreamProcessor health check is based on a health tick which checks if the last successful tick was with in a specified interval. When forcing the rescan of health check by advancing the clock, the streamprocessor is sometimes detected as unhealthy because the last health tick was before the clock was advanced artificially. So it is not a good idea to advance the clock to rescan the health check. Instead we just wait for one healt check interval until the broker is healthy.