test(qa): rename qa base package
Renames the QA package from io.camunda.zeebe.broker.it to
io.camunda.zeebe.it. This better reflects that these are the integration
tests for both the gateway and broker, and it also fixes a problem with
Spring Boot integration tests where the framework scans the whole
io.camunda.zeebe.broker package, which previously included the test
package as well, causing failures due to duplicate beans.