feat(broker): add fixed partition distribution scheme
Adds a new partitioning scheme which allows a user to manually
distribute partitions over the set of nodes in the cluster, optionally
specifying the priority for each node.

The partitioning scheme can now be configured as an experimental
configuration parameter, which by default is ROUND_ROBIN, i.e. the
pre-existing partitioning scheme.

There is minimal validation performed on the new distribution scheme: if
fixed, the broker will not start unless the following conditions are
met:

  - All partitions are configured
  - All partitions are fully replicated