feat(test-util): add logs dumper for container tests
Adds a junit 5 extension which will dump all the logs of a set of
containers on test failure. This can be useful when running container
tests, especially clustered tests where debugging is not possible. The
extension must be registered programatically as the set of containers is
not known at initialization time, and it supports a Supplier as input
instead of just a plain map to support asynchronous cases. See the
Javadoc on how to use it.